+++ src/DomainCommonExtensions/CommonExtensions/Encryption/AESEncryptionExtensions.cs
@@ -36,18 +36,24 @@ namespace DomainCommonExtensions.CommonExtensions
         /// <param name="plainText">Plaint text to be encrypted.</param>
         /// <param name="key">Encryption key.</param>
         /// <param name="iv">Initialization vector.</param>
+        /// <param name="isLongStringKey">Indicate if is provided by user long key or not. Default value = false.</param>
+        /// <param name="keySize">Key size. Default value = 256.</param>
+        /// <param name="blockSize">Block size. Default value = 128.</param>
         /// <returns></returns>
-        public static string AesEncryptString(this string plainText, string key, byte[] iv)
+        public static string AesEncryptString(this string plainText, string key, byte[] iv, bool isLongStringKey = false, int keySize = 256, int blockSize = 128)
         {
             key.ThrowIfArgNullOrEmpty(nameof(key));
             plainText.ThrowIfArgNull(nameof(plainText));
             iv.ThrowIfArgNull(nameof(iv));
 
-            //var iv = new byte[16];
             byte[] array;
             using (var aes = Aes.Create())
             {
-                aes.Key = Encoding.UTF8.GetBytes(key);
+                aes.KeySize = keySize;
+                aes.BlockSize = blockSize;
+                aes.Padding = PaddingMode.PKCS7;
+                aes.Mode = CipherMode.CBC;
+                aes.Key = isLongStringKey.IsTrue() ? GetHashByteKey(key) : Encoding.UTF8.GetBytes(key);
                 aes.IV = iv;
                 var encryptor = aes.CreateEncryptor(aes.Key, aes.IV);
                 using (var memoryStream = new MemoryStream())
@@ -73,18 +79,24 @@ namespace DomainCommonExtensions.CommonExtensions
         /// <param name="cipherText">Encrypted text to be decrypted</param>
         /// <param name="key">Decryption key.</param>
         /// <param name="iv">Initialization vector.</param>
+        /// <param name="isLongStringKey">Indicate if is provided by user long key or not. Default value = false.</param>
+        /// <param name="keySize">Key size. Default value = 256.</param>
+        /// <param name="blockSize">Block size. Default value = 128.</param>
         /// <returns></returns>
-        public static string AesDecryptString(this string cipherText, string key, byte[] iv)
+        public static string AesDecryptString(this string cipherText, string key, byte[] iv, bool isLongStringKey = false, int keySize = 256, int blockSize = 128)
         {
             key.ThrowIfArgNullOrEmpty(nameof(key));
             cipherText.ThrowIfArgNullOrEmpty(nameof(cipherText));
             iv.ThrowIfArgNull(nameof(iv));
 
-            //var iv = new byte[16];
             var buffer = Convert.FromBase64String(cipherText);
             using (var aes = Aes.Create())
             {
-                aes.Key = Encoding.UTF8.GetBytes(key);
+                aes.KeySize = keySize;
+                aes.BlockSize = blockSize;
+                aes.Padding = PaddingMode.PKCS7;
+                aes.Mode = CipherMode.CBC;
+                aes.Key = isLongStringKey.IsTrue() ? GetHashByteKey(key) : Encoding.UTF8.GetBytes(key);
                 aes.IV = iv;
                 var decryptor = aes.CreateDecryptor(aes.Key, aes.IV);
                 using (var memoryStream = new MemoryStream(buffer))
@@ -98,6 +110,19 @@ namespace DomainCommonExtensions.CommonExtensions
                     }
                 }
             }
+        }
+
+        /// <summary>
+        ///     Get AES key byte from user key hash.
+        /// </summary>
+        /// <param name="userKey">User custom long key.</param>
+        /// <returns></returns>
+        private static byte[] GetHashByteKey(string userKey)
+        {
+            var keyBytes = Encoding.UTF8.GetBytes(userKey);
+            using var md5 = MD5.Create();
+
+            return md5.ComputeHash(keyBytes);
         }
     }
 }